refactor: move GlobalContext to file
